// src/app/model/user.ts
export interface User {
        id: number;
        f_name: string;
        l_name: string;
        email: string;
        dept: string;
        designation: string;
        phone: string;
        password: string;
        is_active: number;
        type_id: number;
        file: File;
        course_no:string;
        pers_no:string;
        appt:string;
        user_image: string;
        updated_at: any;
        created_at: any;
}